+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/source/Endpoint.java
@@ -79,6 +79,9 @@ public class Endpoint extends Source {
     @Getter
     @Setter
     private List<String> tags;
+    @Getter
+    @Setter
+    private SideCar sideCar = new SideCar();
 
     @Override
     public void prepare() {