+++ oap-server/server-receiver-plugin/skywalking-mesh-receiver-plugin/src/test/java/org/apache/skywalking/oap/server/receiver/mesh/MeshDataMock.java
@@ -53,6 +53,7 @@ public class MeshDataMock {
                                                  .setStatus(true)
                                                  .setProtocol(Protocol.HTTP)
                                                  .setDetectPoint(DetectPoint.server)
+                                                 .setInternalErrorCode("rate_limited")
                                                  .build());
         }
         meshObserver.onCompleted();